pyproject.toml
[tool.poetry]
name = "api-service"
version = "0.1.0"
description = ""
authors = ["Igor Andrade <igor.andrade@mail.uft.edu.br>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.27.2"
django-cors-headers = "4.3.1"
pytest-django = "4.7.0"
django = "4.2.13"
djangorestframework = "^3.15.1"
mysqlclient = "^2.2.4"
celery = "^5.4.0"
redis = "^5.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
in-project = true